Made frontend responsive
Used Twitter Bootstrap library in replacement of 960 to get responsive
layout.
The menus in the sub pages are a work in progress. Ideally all user
account stuff should be under a separate user menu.
As a side effect of twitter bootstrap some of the styles have changed
fonts, forms, buttons, alerts etc. I think it's for the better.
Also fixed a bug with length being a required field in the edit details
form but it being commented out for the time being.
In the spirit of the cleanup I've also added a background image across
the header. Not exactly a requirement for responsive layout.