Update members page payment history
I've moved the payment history table to show regardless of whether a user is or is not currently a member. As someone who was previously a member I'd like to see what I paid, even if I am no longer a member (and am unlikely to be in the future).